% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_save_extract.R
\name{admin.save_extract}
\alias{admin.save_extract}
\title{Saves an HMIS CSV/XML Export in .rda file format.}
\usage{
admin.save_extract(extractPath, extractDate, saveInDirectory, full = FALSE)
}
\arguments{
\item{extractPath}{the path to the folder containing the HMIS CSV/XML Export .csv files.}

\item{extractDate}{the date the CSV/XML extract was pulled from HMIS: YYYY-MM-DD.}

\item{saveInDirectory}{the directory in which you want to save the output file.}

\item{full}{Defaults to FALSE. If set to TRUE it will include all 14 CSV files from the XML extract instead of the usual 9.}
}
\value{
saves a file called hmisData_extractdate.rda with 9 tibbles: client, services, entry, exit, project, healthanddv, incomebenefits, organization, and disabilities.
}
\description{
Uses the hmis.import_extract function to read in .csv files from data exported from HMIS per the specified date and saves each tibble to an hmisData_extractdate.rda file instead of returning a list of tibbles. NOTE: This requires you to have a RStudio Project open and use the here() package!
}
