% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmis_estimate_chronicity_list.R
\name{hmis.estimate_chronicity_list}
\alias{hmis.estimate_chronicity_list}
\title{Chronicity Calculator for a Data Frame of Clients}
\usage{
hmis.estimate_chronicity_list(entry_file, extractDate)
}
\arguments{
\item{entry_file}{the 'entry' data frame from the list generated/returned by the import_hmis function.}

\item{extractDate}{the date the CSV/XML extract was pulled from HMIS: YYYY-MM-DD.}
}
\value{
TRUE or FALSE.
}
\description{
Estimates whether a client may be experiencing chronic homelessness per the available information in the HMIS export.
}
