% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmis_import_extract.R
\name{hmis.import_extract}
\alias{hmis.import_extract}
\title{Read in an HMIS CSV/XML Export}
\usage{
hmis.import_extract(
  extractPath,
  extractDate,
  ...,
  include_disabilities = FALSE
)
}
\arguments{
\item{extractPath}{the path to the folder containing the HMIS CSV/XML Export .csv files.}

\item{include_disabilities}{OPTIONAL: specifies whether to import the disabilities file (defaults to FALSE).}
}
\value{
a list of 8-9 tibbles: client, services, entry, exit, project, healthanddv, incomebenefits, & organization (and disabilities if include_disabilities = TRUE).
}
\description{
Reads in .csv files from data exported from HMIS per the specified date and returns a list of tibbles.
}
