% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut.R
\name{shortcut}
\alias{shortcut}
\title{Quickly Navigates to Certain Folders in the Research-and-Evaluation-Team Filesystem}
\usage{
shortcut(bookmarkedDirectory, ...)
}
\arguments{
\item{bookmarkedDirectory}{as a "string", the abbreviated name of the folder your file is in: "extracts," "bnl," "db inputs," or "db outputs."}

\item{...}{the rest of the path to you file as you would type it in the here() function.}
}
\value{
Returns the first part of the set of strings to use in the here() function to reach your destination folder, so for example you can type shortcut("extracts", "hmisData_XXXX-XX-XX.rda") instead of here("../", "../", "Datasets", "HMIS_Extracts", "hmisData_XXXX-XX-XX.rda").
}
\description{
Shortcut() is a quicker way to use the here() function, customized to provide quick access to the HMIS extracts, CE BNL, and Dashboard Input/Output folders from the Small Data Requests, Dashboards, and Needs and Gaps folders.
}
